Made the input dialog always-on
The dialog is now always shown, and has a size of zero instead of being
fullscreen & transparent. The dialog is given focus whenever the scan
button is pressed.

The plugin now also sends back `null` to the JavaScript code if the scan
didn't return any input, instead of just silencing it.

Fixes #1 (Scan event appears to run twice)